Teacher is now displaying the table, todo:edit and insert improvements

+++ Model/TeacherLoader.php
@@ -8,11 +8,23 @@ class TeacherLoader
     public function loadTeachers()
     {
         $pdo = Connection::openConnection();
-        $handle = $pdo->prepare('SELECT t.id as id, t.Name as name, c.id as classId, c.name as className t.email as email FROM teacher t LEFT JOIN class c on t.id=c.teacherId');
+        $handle = $pdo->prepare('SELECT t.id as id, t.Name as name, c.id as classId, c.name as className, t.email as email FROM teacher t LEFT JOIN class c on t.id=c.teacherId');
         $handle->execute();
         $teachers = $handle->fetchAll();
         foreach ($teachers as $t) {
-            $this->teachers[] = new Teacher($t['name'], $t['id'], $t['email'], $t['classId'], $t['className']);
+            $handle = $pdo->prepare('SELECT name, id FROM student WHERE teacherId = :id');
+            $handle->bindValue(':id', $t['id']);
+            $handle->execute();
+            $students = $handle->fetchAll();
+            $temp = [];
+            foreach ($students as $s) {
+                $student = array("name" => $s['name'], "id" => $s['id']);
+                $temp[] = $student;
+            }
+            var_dump($t['classId']);
+            $teacherobj = new Teacher($t['name'], $t['id'], $t['email'], $t['classId'], $t['className']);
+            $teacherobj->setStudents($temp);
+            $this->teachers[] = $teacherobj;
         }
 
     }
@@ -43,12 +55,19 @@ class TeacherLoader
         $handle->execute();
     }
 
-    public static function deleteTeacher( $id)
+    public static function deleteTeacher($id)
     {
         $pdo = Connection::openConnection();
-        $handle = $pdo->prepare('DELETE FROM teacher WHERE id = :id');
-        $handle->bindValue(':id', $id);
+        $handle = $pdo->prepare('SELECT * FROM class c LEFT JOIN student s USING (teacherId) WHERE c.teacherId = :teacherId');
+        $handle->bindValue(':teacherId', $id);
         $handle->execute();
+        $foundInClasses = $handle->fetchAll();
+        if (empty($foundInClasses)) {
+            $pdo = Connection::openConnection();
+            $handle = $pdo->prepare('DELETE FROM teacher WHERE id = :id');
+            $handle->bindValue(':id', $id);
+            $handle->execute();
+        }
     }
 
     public function selectTeacher($id)
@@ -58,5 +77,10 @@ class TeacherLoader
                 return $teacher;
             }
         }
+    }
+
+    public function studentsForTeacher()
+    {
+
     }
 }
